rename benchmark example to avoid collisions with google's benchmark lib
