+++ src/examples/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(aes128)
-add_subdirectory(benchmark)
 add_subdirectory(benchmark_integers)
+add_subdirectory(benchmark_primitive_operations)
 add_subdirectory(benchmark_providers)
 add_subdirectory(example_template)
 add_subdirectory(sha256)
